test: test entire rate limiter

+++ throttled/exceptions.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 
 class RateLimitExceeded(Exception):
-    def __init__(self, key: str, retry_after: Optional[int] = None):
+    def __init__(self, key: str, retry_after: Optional[float] = None):
         self.key = key
         self.retry_after = retry_after
         super().__init__()
